ControllerService: Always attach volumes in read-write
It's safer to always attach volumes in read-write mode. The volumes
themselves might be mounted as readonly if read-only access is
requested. The driver does not support MULTI_NODE_READER_ONLY, thus
it's safer to always attach the underlying block device as read-write
so that a filesystem mounted as read-only would be able to replay
its journal.

Change-Id: I4c6b712a76be2254fa4a789229dd61be5a8bc985